# public/openapi.yaml
openapi: 3.0.3
info:
  title: Fine-Tune Labs API
  version: 1.0.0
  description: |

    # Fine-Tune Labs Public API

    Production-grade API for LLMOps, model training, and monitoring.

    ## Features

    - 🎯 **Widget Apps** - Deploy production monitoring widgets
    - 📊 **Analytics & Metrics** - Real-time model performance tracking
    - 🚀 **Inference Deployment** - Deploy and manage model endpoints
    - 🤖 **Model Management** - CRUD operations for LLM models
    - 📈 **Training Jobs** - Monitor fine-tuning progress

    ## Authentication

    All endpoints require Bearer token authentication:

    ```
    Authorization: Bearer YOUR_API_TOKEN
    ```

    Get your API token from the Fine-Tune Labs dashboard.

    ## Base URL

    - **Development**: `http://localhost:3000`
    - **Production**: `https://finetunelab.ai`

    ## Rate Limits

    - 1000 requests per minute per API key
    - 10,000 requests per day per API key

    ## Support

    - Documentation: https://finetunelab.ai/docs
    - Email: support@finetunelab.ai
  contact:
    name: Fine-Tune Labs API Support
    email: support@finetunelab.ai
    url: https://finetunelab.ai/support
  license:
    name: Proprietary
    url: https://finetunelab.ai/terms
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://finetunelab.ai
    description: Production server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from Supabase authentication
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
        details:
          type: string
          description: Additional error details
        code:
          type: string
          description: Error code
      required:
        - error
...
security:
  - bearerAuth: []
...
paths:
...
  /api/analytics/traces:
    get:
      summary: Get execution traces
      description: |
        Retrieve detailed execution traces of LLM operations for debugging and analysis.

        Traces provide hierarchical view of:
        - Request/response cycles
        - Tool calls and their execution
        - Model invocations
        - Performance timing data
        - Error information

        **Use Cases:**
        - Debug production issues
        - Performance profiling
        - Track multi-step agent operations
        - Audit model behavior
      tags:
        - Analytics
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: conversation_id
          schema:
            type: string
          description: Filter by conversation ID
        - in: query
          name: limit
          schema:
            type: integer
            default: 100
          description: Maximum traces to return
        - in: query
          name: start_date
          schema:
            type: string
            format: date-time
          description: Filter traces after this date
      responses:
        '200':
          description: Traces retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  traces:
                    type: array
                    items:
                      type: object
                      properties:
                        trace_id:
                          type: string
                        span_name:
                          type: string
                          example: llm.completion
                        duration_ms:
                          type: number
                          example: 1234.56
                        model_name:
                          type: string
                          example: gpt-4-turbo
                        operation_type:
                          type: string
                          example: completion
                        children:
                          type: array
                          description: Nested child traces
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      summary: Capture execution trace
      description: |
        Send a trace event for an LLM operation. Used to track and debug production systems.

        Typically called automatically by your application code when integrated with
        tracing middleware.
      tags:
        - Analytics
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - trace_id
                - span_id
                - span_name
                - start_time
              properties:
                trace_id:
                  type: string
                  example: trace_abc123
                span_id:
                  type: string
                  example: span_xyz789
                span_name:
                  type: string
                  example: llm.completion
                start_time:
                  type: string
                  format: date-time
                end_time:
                  type: string
                  format: date-time
                duration_ms:
                  type: number
                operation_type:
                  type: string
                  example: completion
                model_name:
                  type: string
                  example: gpt-4-turbo
                metadata:
                  type: object
      responses:
        '201':
          description: Trace captured successfully
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...